fix(indices): ограничить цену индекса целыми рублями
- поле цены принимает только целые рубли за тонну и форматирует тысячи

- API отклоняет дробные значения цены

- добавлены тесты UI-helper и zod-схемы

+++ apps/api/src/content/content.schemas.ts
@@ -83,7 +83,7 @@ export const priceIndexInputSchema = z.object({
 
 export const priceIndexValueInputSchema = z.object({
   date: z.string().datetime(),
-  price: z.number().positive(),
+  price: z.number().int().positive(),
 });
 
 export const learningModuleInputSchema = z.object({
